+++ test/td_cache/field_cache_test.exs
@@ -81,12 +81,14 @@ defmodule TdCache.FieldCacheTest do
     end
 
     test "reads an external id" do
-      [system, group, name, field] = ["Test System", "Test Group", "Test Structure", "Test Field"]
-      external_id = Enum.join([system, group, name, field], ".")
-      assert {:ok, 1} = Redix.command(["SADD", "data_fields:external_ids", external_id])
-      assert FieldCache.get_external_id(system, group, name, field) == external_id
-      assert FieldCache.get_external_id(system, group, name, "foo") == nil
-      Redix.command(["SREM", "data_fields:external_ids", external_id])
+      system_external_id = "Test System"
+      [group, name, field] = ["Test Group", "Test Structure", "Test Field"]
+      external_id = Enum.join([group, name, field], ".")
+      assert {:ok, 1} = Redix.command(["SADD", "structures:external_ids:#{system_external_id}", external_id])
+      assert FieldCache.get_external_id(system_external_id, external_id) == external_id
+      unexisting_external_id = Enum.join([group, name, "foo"], ".")
+      assert FieldCache.get_external_id(system_external_id, unexisting_external_id) == nil
+      assert {:ok, 1} = Redix.command(["SREM", "structures:external_ids:#{system_external_id}", external_id])
     end
   end
 end