+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Changed
+
+- [TD-1970] Changed data_fields:external_ids set for structures:external_ids:{system_external_id} set
+
 ## [3.0.5] 2019-06-27
 
 ### Added